use Wallet instead of BaseWallet
